Add support for plugins with ViewHierarchy
The goal here is for the developer to be able to dump custom
information about their views when generating screenshots. A simple
example as shown in the test might be store getText() if the view is a
TextView.

At this point there is no external entry point to add the plugin yet.