Add Windows support
This change attempts to make the plugin's python scripts work on
Windows. Right now, the only thing we check for is that the tests
pass on Windows (in particular Python 3 on Windows 10).

There are many more possible configurations, but it's going to be hard
to check all of those, so I'm going to hope the community steps in to
verify that this works.