Spruce up Results Page
The screenshot test result page looked a little sad / bland before, this
makes it look better (tm). I also inlined the view hierarchy since the
previous button to view it didnt work, and it didnt make much sense not
to. I also plan on converting the XML into a collapsable hierarchy at
some point, but this is a quick start on a redesign of the page