Add ability to run screenshot tests on multiple devices at once
Summary:
Issue: #112

The objective for this PR is to be able to run screenshot tests in multiple devices.

I have added a new python class called DeviceNameCalculator which is intended to calculate a unique device identifier. This is calculated using some adb commands. This device identifier will be the directory where the screenshots will be saved.

I have thought that a reliable device identifier should be based on:
- Android API Level
- Google Play Services
- Screen density
- Screen height and width

For example, a device with API level 25, Google Play Services, xxhdpi as screen density and a screen of 1920x1600 the name results as API_25_GP_XXHDPI_1920_1600. This will be the folder inside screenshots folder where the screenshots will be saved.

Screenshot filename will be the same as usual.

For testing these new changes you can use app-example-androidjunitrunner project from examples folder.

Users should clear screenshot folder and run recordMode in order to create new folders based on the device which they are run.
Closes https://github.com/facebook/screenshot-tests-for-android/pull/115

Reviewed By: sjkirby

Differential Revision: D6128610

Pulled By: xiphirx

fbshipit-source-id: 2c251da62bb1bf24084327c462856d6384a2a807