Move sample apps to root project level
Summary:
It was annoying to manage 5 separate gradle installations / configs.
Now each sample app is at the root project level, with one gradle installation.

This makes testing and other things much easier. It also removes a ton of cruft.

I plan on consolidating the sample apps later on, as there is no good reason to have 4.

Reviewed By: nscoding

Differential Revision: D6653215

fbshipit-source-id: 2c09d00abab67519263c4891082fd124ffefe553