Remove buck dependencies on test code
Summary:
The library itself does not require any dependencies on junit or dexmaker. They seem to be needed only in test code.
This diff removes those dependencies from the BUCK file. I keep the definitions in SCREENSHOT_TESTS_DEFS though because one day we might add a buck module to run the tests.

Reviewed By: xiphirx

Differential Revision: D7145395

fbshipit-source-id: 3fbc740101f9397ef714b9cfc94bf02f43ad77a9