Ensure InputMethodManager is available
Summary: On older versions of Android, if a view, like EditText, requests focus prior to being attached to the window, then we hit a NPE since the singleton instance hasn't been created yet. This works around that by causing the instance to be created before we dispatch the window attachment.

Reviewed By: sjkirby

Differential Revision: D7783889

fbshipit-source-id: bec58e4db4055f1059c3096b4556ee29418ff572