Make tests pass (#201)
Summary:
Building the plug-in and running the tests by following the instructions in the README (`gradle :plugin:pyTests` and `gradle :core:connectedAndroidTest`) is currently not possible because of a broken test and a missing dependency. This pull request addresses both issues and builds at least on my machine (MacOS X 10.13). There isn't a publicly available CI server that I somehow didn't spot?
Pull Request resolved: https://github.com/facebook/screenshot-tests-for-android/pull/201

Differential Revision: D13454751

Pulled By: xiphirx

fbshipit-source-id: f8c6761828debe6b004058f00d6b93c221e297bc