Add the concept of a version to the metadata JSON file
Summary:
Adding a version to the metadata file will allow us to iterate on the structure of the hierarchies themselves without worrying about making all tests break when we do so.

As of D15403508 and D15403508 will make sure that the hierarchy check is considered a match if the version changes.

Reviewed By: sjkirby

Differential Revision: D15404142

fbshipit-source-id: 43d037819d2359b2bd8a4a7739654ae3f5e1eb08