Add AccessibilityUtil class with a data structure for the accessibility tree
Summary:
This is the first diff in a stack of N that will add the concept of accessibility evaluations to the screenshot testing framework.  This diff does a bit of prep work and cleanup by adding a new AccessibilityUtil class that generates the accessibility tree as class of its own, rather than creating the tree in JSON itself.

Subsequent diffs will use this AXTreeNode class to run evaluations over the tree to find accessibility problems.

Reviewed By: sjkirby

Differential Revision: D15251083

fbshipit-source-id: 46104d86b5f426a1c0bf0adea5e1e4551dc58ea2