Force Litho to render in accessibility mode, and bump the metadata version number
Summary:
This was originally added in D13287519 but was reverted and never re-added.

This property forces Litho to render its view hierarchy as if an accessibility service was running, which is necessary for the actual accessibility hierarchy generation to be accurate to what real users will see.

This change will change the actual hierarchies for all existing tests however, so we must also bump the metadata version number at the same time so as not to cause a massive amount of changes to be detected.

Reviewed By: sjkirby

Differential Revision: D15443923

fbshipit-source-id: 243fc42b94d0a809bc7620aac23321d7df6f1695