Add accessibility evaluation to detect focusable views without any text to announce.
Summary:
This adds the first accessibility evaluation to the screenshot test framework.  This evaluation checks for views that are focusable by Talkback, but have no text to speak aloud.

These views will generally be leaf nodes that have no focusable ancestors, no label of their own, and no children that have text or a contentDescription.

The `axIssues` key in the dump file will hold the results of all evaluations, and each evaluation will have the following structure:

```
  "axIssues": [
    {
      "id": "id_of_evaluation",
      "name": "Name of the Evaluation",
      "description": "Short description of the evaluation, explaining its purpose and what the results mean.",
      "elements": [
        {
          "name": "Button",
          "class": "com.android.widget.Button",
          "position": {
            "left": 0,
            "top": 0,
            "width": 300,
            "height": 300
          },
          "suggestions": [
            "a suggested fix for the issue."
          ]
        }
      ]
    },
  ]
```

The `suggestions` array on each element contains a list of one or more suggestions on how to resolve the issue.  These can be contextual with specific suggestions for different element types or situations.

These issues are not yet surfaced in the screenshot report page, but are being written to the JSON file for each test.

Reviewed By: sjkirby

Differential Revision: D15405075

fbshipit-source-id: 207754e083766185d924b0aa11dec551d7f64e66