Remove debug specific configurations from `ComponentsConfiguration`
Summary:
## Important
This is a "copy" of D52563388, which was reverted due to three classes that were still referring to the old configurations and somehow Sandcastle missed. I'm updating those references on top of this copy.

## Original Description
This diff removes debug specific configurations from `ComponentsConfiguration` in Litho into a new abstraction named `LithoDebugConfigurations`. The goal is that we move further into having `ComponentsConfiguration` to be more focused in internals/behaviors of Litho.

Reviewed By: adityasharat

Differential Revision: D52779201

fbshipit-source-id: 8f540efbe2691f7b8c22d1b199a14ff02f986c9c