Bulk update of unwritten private fields to final/val
Summary:
drop-conflicts

Adds non-mutability enforcement to private fields that we've detected are never written to. The entire mod is adding the final specifier for java and converting var -> val for kotlin.

# Methodology for this local mod:
I was using this as a test for the diff2input workflow for diff2codemod, that WIP prompt generated this input script D81237144.

From there I just had devmate blast out some regexes in a script (P1921928763) and fed that those targets (and variables).

#  Why this should be pretty safe
I've personally eyeballed each line in these diffs to avoid clownyness but feel free to take a second pass, more eyes are appreciated.

Since this is a bytecode compile time check: if it builds, it ships.

The only (minor) chance of a breakage from this is a bizarre reflection usage or more likely a unit test failure due to similar behavior (cough VisibleForTesting). I've omitted modifying any fields that were explicitly annotated as VisibleForTesting for now, some may be fine (i.e. read only) but those can be addressed in a followup pass.

Reviewed By: ianlevesque

Differential Revision: D81078997

fbshipit-source-id: 6f5c5ea67e14f79e5bc33ef0c09dc210272d0c3a